refactored reporting/base test cases to mitigate expensive fixture loads
