+++ tests/base.py
@@ -1,7 +1,6 @@
 import os
 from splice.environment import Environment
 from splice.webapp import create_webapp
-from datetime import datetime
 env = Environment.instance(test=True)
 
 from flask.ext.testing import TestCase
@@ -21,13 +20,6 @@ class BaseTestCase(TestCase):
         self.create_app()
         self.env.db.create_all()
 
-        def values(fd):
-            for line in fd:
-                row = line.split(',')
-                # sqlalchemy doesn't like date strings....
-                row[1] = datetime.strptime(row[1], "%Y-%m-%d")
-                yield row
-
         def tile_values(fd):
             for line in fd:
                 row = line.split(',')
@@ -35,14 +27,7 @@ class BaseTestCase(TestCase):
                     ('id', 'target_url', 'bg_color', 'title', 'type', 'image_uri', 'enhanced_image_uri', 'locale'),
                     row))
 
-        # load db
-        from splice.models import impression_stats_daily, Tile
-        conn = Environment.instance().db.engine.connect()
-        with open(self.get_fixture_path('impression_stats.csv')) as fd:
-            for row in values(fd):
-                ins = impression_stats_daily.insert().values(row)
-                conn.execute(ins)
-
+        from splice.models import Tile
         session = Environment.instance().db.session
         with open(self.get_fixture_path('tiles.csv')) as fd:
             for row in tile_values(fd):
